requirements/base.txt
Django==3.2.9
python-decouple==3.5
django-environ==0.8.1
django-model-utils==4.2.0
django-cleanup==5.2.0
django-import-export==2.6.1
django-cors-headers==3.10.0
djangorestframework==3.12.4
dj-rest-auth==2.1.11
djangorestframework-simplejwt==5.0.0
djangorestframework-filters==1.0.0.dev0
drf-renderer-xlsx==0.4.3
django-allauth==0.46.0
PyJWT==2.3.0
Scrapy==2.5.0
scrapy_djangoitem==1.1.1
scrapyd-api==0.0.1
django_celery_results==2.2.0
sentry-sdk==1.4.3
django-grappelli==2.15.3
django_admin_rangefilter==0.8.2